Add code to sniff the line ending of the CSV file
Calling code can optionally specify the line ending. If they don't specify the line ending then it will be sniffed automatically. By passing the correct line ending to the streamReader() no line ending characters are included in the lines returned from the streamReader. Note that any files with mixed line endings will result in undefined behavior. But it was always like that.